# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ExportFormat": {
        "description": "Export formats (`REQ-EXP-001`, `REQ-EXP-002`).",
        "enum": [
          "pdf",
          "pptx"
        ],
        "title": "ExportFormat",
        "type": "string"
      },
      "ExportOut": {
        "description": "One export and where it stands (`REQ-EXP-006`, `REQ-EXP-007 AC-2`).",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "download_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Download Url"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "format": {
            "$ref": "#/components/schemas/ExportFormat"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "size_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Size Bytes"
          },
          "status": {
            "$ref": "#/components/schemas/ExportStatus"
          },
          "theme": {
            "$ref": "#/components/schemas/ExportTheme"
          },
          "version_number": {
            "title": "Version Number",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "version_number",
          "format",
          "theme",
          "status",
          "error",
          "size_bytes",
          "created_at",
          "completed_at"
        ],
        "title": "ExportOut",
        "type": "object"
      },
      "ExportRequest": {
        "description": "Which format, in which of the six themes (`REQ-EXP-003`).",
        "properties": {
          "format": {
            "$ref": "#/components/schemas/ExportFormat"
          },
          "theme": {
            "$ref": "#/components/schemas/ExportTheme"
          }
        },
        "required": [
          "format",
          "theme"
        ],
        "title": "ExportRequest",
        "type": "object"
      },
      "ExportStatus": {
        "description": "Export job lifecycle. Retryable without re-running research (`NFR-REL-003`).",
        "enum": [
          "pending",
          "running",
          "ready",
          "failed"
        ],
        "title": "ExportStatus",
        "type": "string"
      },
      "ExportTheme": {
        "description": "The six predefined themes (`REQ-EXP-003`). Visual definitions are `OPEN-22`.",
        "enum": [
          "professional",
          "investor",
          "modern",
          "corporate",
          "minimal",
          "dark"
        ],
        "title": "ExportTheme",
        "type": "string"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Turns a question into an evidence-backed report with sources, analysis, charts and follow-up.",
    "title": "ScrapR",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/exports/{export_id}": {
      "get": {
        "description": "One export; with a short-lived download link once it is ready.",
        "operationId": "get_export_v1_exports__export_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Export Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_account",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Account"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Export",
        "tags": [
          "exports"
        ]
      }
    },
    "/v1/exports/{export_id}/retry": {
      "post": {
        "description": "Render a failed export again, without re-running research (`NFR-REL-003`).",
        "operationId": "retry_export_v1_exports__export_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "export_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Export Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_account",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Account"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Session"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Export",
        "tags": [
          "exports"
        ]
      }
    },
...
    "/v1/research/{session_id}/versions/{version_number}/exports": {
      "get": {
        "description": "Every export of one version, newest first (export version tracking).",
        "operationId": "list_exports_v1_research__session_id__versions__version_number__exports_get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_number",
            "required": true,
            "schema": {
              "title": "Version Number",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_account",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Account"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExportOut"
                  },
                  "title": "Response List Exports V1 Research  Session Id  Versions  Version Number  Exports Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Exports",
        "tags": [
          "exports"
        ]
      },
      "post": {
        "description": "Queue a PDF or PowerPoint of one version in one theme.\n\nAsking again for the same version, format and theme returns the export\nalready queued or ready rather than rendering it twice: the version cannot\nchange, so neither can the file.",
        "operationId": "request_export_v1_research__session_id__versions__version_number__exports_post",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Session Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_number",
            "required": true,
            "schema": {
              "title": "Version Number",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_account",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Account"
            }
          },
          {
            "in": "cookie",
            "name": "scrapr_session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Scrapr Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExportRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExportOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request Export",
        "tags": [
          "exports"
        ]
      }
    }
  }
}